fixapp/requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
djangorestframework==3.15.2
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-vision==3.9.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
idna==3.10
numpy==2.2.1
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
